+++ src/flows/process_erikas_tweets.py
@@ -2,7 +2,7 @@ from mlflow import log_metric, log_param, log_artifacts
 import pandas as pd
 from prefect import flow, task, get_run_logger
 from ..common import load_raw_dataset, save_dataset, drop_duplicated_rows
-from .utils import drop_unused_tweet_columns, filter_letters, remove_stopwords, lemmatize, tokenize, get_bag_of_words
+from ..preprocess import drop_unused_tweet_columns, filter_letters, remove_stopwords, lemmatize, tokenize, get_bag_of_words
 
 
 @task
